Move jest configurations to proper files
