+++ test/e2e/route-access/anonymous.e2e-spec.ts
@@ -2,18 +2,18 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from '../../src/auth/auth.module';
+import { AuthModule } from '../../../src/auth/auth.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import databaseConfig from '../../src/config/database.config';
-import securityConfig from '../../src/config/security.config';
-import environmentConfig from '../../src/config/environment.config';
-import integrationsConfig from '../../src/config/integrations.config';
-import mediaConfig from '../../src/config/media.config';
-import { UserModule } from '../../src/user/user.module';
+import databaseConfig from '../../../src/config/database.config';
+import securityConfig from '../../../src/config/security.config';
+import environmentConfig from '../../../src/config/environment.config';
+import integrationsConfig from '../../../src/config/integrations.config';
+import mediaConfig from '../../../src/config/media.config';
+import { UserModule } from '../../../src/user/user.module';
 import { randomUUID } from 'crypto';
-import { TourModule } from '../../src/tour/tour.module';
-import { LookupModule } from '../../src/lookup/lookup.module';
-import { MediaModule } from '../../src/media/media.module';
+import { TourModule } from '../../../src/tour/tour.module';
+import { LookupModule } from '../../../src/lookup/lookup.module';
+import { MediaModule } from '../../../src/media/media.module';
 
 const AUTH_ROUTE_PREFIX = '/auth';
 const USER_ROUTE_PREFIX = '/users';
