+++ test/e2e/route-access/authenticated-user.e2e-spec.ts
@@ -2,22 +2,22 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { ExecutionContext, INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from '../../src/auth/auth.module';
+import { AuthModule } from '../../../src/auth/auth.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import databaseConfig from '../../src/config/database.config';
-import securityConfig from '../../src/config/security.config';
-import environmentConfig from '../../src/config/environment.config';
-import integrationsConfig from '../../src/config/integrations.config';
-import mediaConfig from '../../src/config/media.config';
-import { UserModule } from '../../src/user/user.module';
+import databaseConfig from '../../../src/config/database.config';
+import securityConfig from '../../../src/config/security.config';
+import environmentConfig from '../../../src/config/environment.config';
+import integrationsConfig from '../../../src/config/integrations.config';
+import mediaConfig from '../../../src/config/media.config';
+import { UserModule } from '../../../src/user/user.module';
 import { randomUUID } from 'crypto';
-import { JwtAuthGuard } from '../../src/auth/guards/jwt-auth.guard';
-import { TourModule } from '../../src/tour/tour.module';
-import { LookupModule } from '../../src/lookup/lookup.module';
-import { MediaModule } from '../../src/media/media.module';
-import { AuthenticatedUserDto } from '../../src/user/dto/user.dto';
+import { JwtAuthGuard } from '../../../src/auth/guards/jwt-auth.guard';
+import { TourModule } from '../../../src/tour/tour.module';
+import { LookupModule } from '../../../src/lookup/lookup.module';
+import { MediaModule } from '../../../src/media/media.module';
+import { AuthenticatedUserDto } from '../../../src/user/dto/user.dto';
 import { Repository } from 'typeorm';
-import { Tour } from '../../src/tour/entities/tour.entity';
+import { Tour } from '../../../src/tour/entities/tour.entity';
 import { Seeder } from '../utils/seeder';
 import { EntityCreator } from '../utils/entity-creator';
