+++ test/e2e/utils/seeds.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
-import { User, UserRole } from '../../src/user/entities/user.entity';
-import { Tour } from '../../src/tour/entities/tour.entity';
+import { User, UserRole } from '../../../src/user/entities/user.entity';
+import { Tour } from '../../../src/tour/entities/tour.entity';
 
 export type Seeds = {
   users: User[];
